requirements.txt
astroid==3.0.1
asttokens==2.4.1
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dill==0.3.7
executing==2.0.1
greenlet==3.0.2
idna==3.6
ipython==8.18.1
isort==5.13.1
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tenacity==8.2.3
tokenize-rt==5.2.0
tomlkit==0.12.3
traitlets==5.14.0
types-requests==2.31.0.10
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
